plugin-video: migrate to EditorPlugin
