+++ packages/plugins/video/src/editor.tsx
@@ -1,40 +1,36 @@
 import { OverlayInput } from '@edtr-io/core'
 import { EditorInput, PrimarySettings } from '@edtr-io/editor-ui'
-import { DeprecatedPluginEditorProps } from '@edtr-io/plugin'
 import * as React from 'react'
 
-import { videoState } from '.'
+import { VideoProps } from '.'
 import { VideoRenderer } from './renderer'
 
-export const VideoEditor = (
-  props: DeprecatedPluginEditorProps<typeof videoState> & {
-    renderIntoExtendedSettings?: (children: React.ReactNode) => React.ReactNode
-  }
-) => {
+export const VideoEditor = (props: VideoProps) => {
   const { editable, focused, state } = props
+
+  if (!editable) return <VideoRenderer {...props} />
+
   return (
     <React.Fragment>
       <VideoRenderer {...props} disableCursorEvents={editable} />
-      {props.renderIntoExtendedSettings
-        ? props.renderIntoExtendedSettings(
-            <React.Fragment>
-              <OverlayInput
-                label="Video URL:"
-                value={state.src.value}
-                onChange={e => {
-                  state.src.set(e.target.value)
-                }}
-              />
-              <OverlayInput
-                label="alternativer Text:"
-                value={state.alt.value}
-                onChange={e => {
-                  state.alt.set(e.target.value)
-                }}
-              />
-            </React.Fragment>
-          )
-        : null}
+      {props.renderIntoSettings(
+        <React.Fragment>
+          <OverlayInput
+            label="Video URL:"
+            value={state.src.value}
+            onChange={e => {
+              state.src.set(e.target.value)
+            }}
+          />
+          <OverlayInput
+            label="alternativer Text:"
+            value={state.alt.value}
+            onChange={e => {
+              state.alt.set(e.target.value)
+            }}
+          />
+        </React.Fragment>
+      )}
       {focused ? (
         <PrimarySettings>
           <EditorInput
