+++ packages/private/bundle-size/src/shared/plugins.ts
@@ -17,7 +17,7 @@ import { createScMcExercisePlugin } from '@edtr-io/plugin-sc-mc-exercise'
 import { createSolutionPlugin } from '@edtr-io/plugin-solution'
 import { createSpoilerPlugin } from '@edtr-io/plugin-spoiler'
 import { textPlugin } from '@edtr-io/plugin-text'
-import { videoPlugin } from '@edtr-io/plugin-video'
+import { createVideoPlugin } from '@edtr-io/plugin-video'
 
 const mockUploadFileHandler = (file: File): Promise<UploadedFile> => {
   return readFile(file).then(loaded => {
@@ -49,5 +49,5 @@ export const plugins: Record<
   solution: createSolutionPlugin(),
   spoiler: createSpoilerPlugin(),
   text: textPlugin,
-  video: videoPlugin
+  video: createVideoPlugin()
 }
